Create pptx

+++ R/pptx_server.R
@@ -15,7 +15,6 @@ pptx_server <- function(id) {
       log4r::debug(.le$logger, "pptx_server module started")
       ns <- shiny::NS(id)
       rv <- shiny::reactiveValues(
-        template = "default",  # Default template for the add_images functionality
         uploaded_template = NULL,  # To store the uploaded template for add_images
         uploaded_file = NULL,  # For the sync functionality
         processed_file = NULL
@@ -67,7 +66,7 @@ pptx_server <- function(id) {
           shiny::showModal(shiny::modalDialog(
             title = "Error",
             "No PPTX file was uploaded for syncing. Please try again.",
-            footer = NULL
+            footer = shiny::modalButton("Close")
           ))
           return()
         }
@@ -115,7 +114,7 @@ pptx_server <- function(id) {
           shiny::showModal(shiny::modalDialog(
             title = "Error",
             paste("An error occurred while processing your file:", e$message),
-            footer = NULL
+            footer = shiny::modalButton("Close")
           ))
         })
       })
@@ -141,7 +140,7 @@ pptx_server <- function(id) {
           shiny::showModal(shiny::modalDialog(
             title = "Error",
             "No template file was uploaded. A blank template will be used instead.",
-            footer = NULL
+            footer = shiny::modalButton("Close")
           ))
           return()
         }
@@ -149,7 +148,6 @@ pptx_server <- function(id) {
         # Update the template with the uploaded template for add_images functionality
         log4r::debug(.le$logger, paste0("Template uploaded: ", rv$uploaded_template$datapath))
         log4r::info(.le$logger, "Template uploaded successfully")
-        rv$template <- rv$uploaded_template$datapath
         shiny::removeModal()
       })
 
@@ -176,7 +174,6 @@ pptx_server <- function(id) {
 
         # Reset the template variables
         rv$uploaded_template <- NULL
-        rv$template <- "default"  # Set to the blank template
 
         shiny::removeModal()
 
@@ -212,35 +209,57 @@ pptx_server <- function(id) {
         content = function(file) {
           temp_pptx <- tempfile(fileext = ".pptx")
 
-          remote_url <- gert::git_remote_info()$url
+          remote_url <- gert::git_remote_info()$url # tryCatch in observe as primary
+
+          if (is.null(remote_url)) {
+            shiny::showModal(shiny::modalDialog(
+              title = "Error",
+              "No remote repositories found. Please set a remote repository before using this feature.",
+              footer = NULL
+            ))
+            return()
+          } # Backup for secondary
 
           shiny::showModal(shiny::modalDialog("Creating slides for PowerPoint . . .", footer = NULL))
           log4r::info(.le$logger, "Starting PowerPoint creation process")
           start_time <- Sys.time()
 
           # Use the uploaded template, or the default if not uploaded
-          base_pptx <- if (!is.null(rv$uploaded_template)) rv$uploaded_template$datapath else rv$template
+          base_pptx <- if (!is.null(rv$uploaded_template)) rv$uploaded_template$datapath
 
-          suppressWarnings(create_pptx(
-            remote_url = remote_url,
-            files = selected_items(),
-            output_pptx = temp_pptx,
-            base_pptx = base_pptx
-          ))
+          tryCatch({
+            create_pptx(
+              remote_url = remote_url,
+              files = selected_items(),
+              output_pptx = temp_pptx,
+              base_pptx = base_pptx
+            )
 
-          file.copy(temp_pptx, file)
-          elapsed_time <- Sys.time() - start_time
+            file.copy(temp_pptx, file)
+            elapsed_time <- Sys.time() - start_time
 
-          log4r::info(.le$logger, sprintf("PowerPoint successfully created and downloaded in %.2f seconds.", elapsed_time))
-          on.exit({
-            file.remove(temp_pptx)
+            log4r::info(.le$logger, sprintf("PowerPoint successfully created and downloaded in %.2f seconds.", elapsed_time))
+
             shiny::showModal(shiny::modalDialog(
+              title = "Success",
               sprintf("PowerPoint downloaded successfully in %.2f seconds.", elapsed_time),
+              footer = shiny::modalButton("Close")
             ))
+
+          }, error = function(e) {
+            log4r::error(.le$logger, paste0("Error creating PPTX: ", e$message))
+            shiny::showModal(shiny::modalDialog(
+              title = "Error",
+              paste("An error occurred while creating the PowerPoint:", e$message),
+              footer = shiny::modalButton("Close")
+            ))
+          }, finally = {
+            on.exit({
+              if (file.exists(temp_pptx)) file.remove(temp_pptx)
+            })
           })
         }
       )
-
       output$show_files <- shiny::renderUI({
         htmltools::tagList(
           htmltools::tags$h6("These files reflect the current state of your local repository.
